# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CPCMS \u2014 Course Project Configuration and Management System",
    "description": "Internal University Course Project SCM & Configuration Management System",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_v1_projects__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Update Project",
        "operationId": "update_project_api_v1_projects__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/teams/{team_id}/change-requests": {
      "get": {
        "tags": [
          "Change Requests"
        ],
        "summary": "List Change Requests",
        "operationId": "list_change_requests_api_v1_teams__team_id__change_requests_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ChangeRequestOut"
                  },
                  "title": "Response List Change Requests Api V1 Teams  Team Id  Change Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Create Change Request",
        "operationId": "create_change_request_api_v1_teams__team_id__change_requests_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "team_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Team Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/change-requests/{id}": {
      "get": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Get Change Request",
        "operationId": "get_change_request_api_v1_change_requests__id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Update Change Request",
        "operationId": "update_change_request_api_v1_change_requests__id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeRequestUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/change-requests/{id}/transition": {
      "post": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Transition Change Request",
        "operationId": "transition_change_request_api_v1_change_requests__id__transition_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CRTransitionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChangeRequestOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/change-requests/{id}/impact": {
      "get": {
        "tags": [
          "Change Requests"
        ],
        "summary": "Get Impact Analysis",
        "operationId": "get_impact_analysis_api_v1_change_requests__id__impact_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CRImpactOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CRTransitionRequest": {
        "properties": {
          "to_status": {
            "type": "string",
            "pattern": "^(draft|submitted|under_review|approved|rejected|implemented|verified|closed)$",
            "title": "To Status"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "implemented_commit_sha": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implemented Commit Sha"
          },
          "implemented_version_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Implemented Version Id"
          }
        },
        "type": "object",
        "required": [
          "to_status"
        ],
        "title": "CRTransitionRequest"
      },
      "ChangeRequestCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 250,
            "minLength": 3,
            "title": "Title"
          },
          "description": {
            "type": "string",
            "minLength": 5,
            "title": "Description"
          },
          "reason": {
            "type": "string",
            "minLength": 5,
            "title": "Reason"
          },
          "priority": {
            "type": "string",
            "pattern": "^(low|medium|high|critical)$",
            "title": "Priority",
            "default": "medium"
          },
          "affected_ci_ids": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Affected Ci Ids",
            "default": []
          },
          "components": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Components",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "title",
          "description",
          "reason"
        ],
        "title": "ChangeRequestCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProjectOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "course_id": {
            "type": "string",
            "format": "uuid",
            "title": "Course Id"
          },
          "course_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Code"
          },
          "course_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Course Name"
          },
          "faculty_user_id": {
            "type": "string",
            "format": "uuid",
            "title": "Faculty User Id"
          },
          "faculty_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Faculty Name"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "semester": {
            "type": "string",
            "title": "Semester"
          },
          "academic_year": {
            "type": "string",
            "title": "Academic Year"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "problem_statement": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Problem Statement"
          },
          "objectives": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Objectives"
          },
          "min_team_size": {
            "type": "integer",
            "title": "Min Team Size"
          },
          "max_team_size": {
            "type": "integer",
            "title": "Max Team Size"
          },
          "team_formation_mode": {
            "type": "string",
            "title": "Team Formation Mode"
          },
          "team_formation_deadline": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Formation Deadline"
          },
          "max_projects_per_student": {
            "type": "integer",
            "title": "Max Projects Per Student"
          },
          "allowed_file_exts": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Allowed File Exts"
          },
          "max_file_mb": {
            "type": "integer",
            "title": "Max File Mb"
          },
          "late_policy": {
            "type": "string",
            "title": "Late Policy"
          },
          "late_penalty_percent": {
            "type": "number",
            "title": "Late Penalty Percent"
          },
          "versioning_required": {
            "type": "boolean",
            "title": "Versioning Required"
          },
          "git_required": {
            "type": "boolean",
            "title": "Git Required"
          },
          "baseline_frequency": {
            "type": "string",
            "title": "Baseline Frequency"
          },
          "change_request_required": {
            "type": "boolean",
            "title": "Change Request Required"
          },
          "approval_required": {
            "type": "boolean",
            "title": "Approval Required"
          },
          "branching_policy": {
            "type": "string",
            "title": "Branching Policy"
          },
          "total_marks": {
            "type": "integer",
            "title": "Total Marks"
          },
          "grade_bands": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Grade Bands"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "themes": {
            "items": {
              "$ref": "#/components/schemas/ThemeOut"
            },
            "type": "array",
            "title": "Themes",
            "default": []
          },
          "outcomes": {
            "items": {
              "$ref": "#/components/schemas/OutcomeOut"
            },
            "type": "array",
            "title": "Outcomes",
            "default": []
          },
          "ci_templates": {
            "items": {
              "$ref": "#/components/schemas/CITemplateOut"
            },
            "type": "array",
            "title": "Ci Templates",
            "default": []
          },
          "deadlines": {
            "items": {
              "$ref": "#/components/schemas/DeadlineOut"
            },
            "type": "array",
            "title": "Deadlines",
            "default": []
          },
          "evaluation_criteria": {
            "items": {
              "$ref": "#/components/schemas/CriteriaOut"
            },
            "type": "array",
            "title": "Evaluation Criteria",
            "default": []
          },
          "user_team_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Id"
          },
          "user_team_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Number"
          },
          "user_team_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Team Status"
          }
        },
        "type": "object",
        "required": [
          "id",
          "course_id",
          "faculty_user_id",
          "name",
          "semester",
          "academic_year",
          "min_team_size",
          "max_team_size",
          "team_formation_mode",
          "max_projects_per_student",
          "allowed_file_exts",
          "max_file_mb",
          "late_policy",
          "late_penalty_percent",
          "versioning_required",
          "git_required",
          "baseline_frequency",
          "change_request_required",
          "approval_required",
          "branching_policy",
          "total_marks",
          "grade_bands",
          "status",
          "created_at"
        ],
        "title": "ProjectOut"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}